Rename LocalFileLocator to AppDetails
