+++ app/src/main/java/tech/ula/utils/AndroidUtility.kt
@@ -70,38 +70,10 @@ class BuildWrapper {
 }
 
 class ConnectionUtility {
-    @Throws(Exception::class)
     fun getUrlInputStream(url: String): InputStream {
         val conn = URL(url).openConnection() as HttpURLConnection
         conn.requestMethod = "GET"
         return conn.inputStream
-    }
-}
-
-// TODO move to UlaFiles?
-class LocalFileLocator(private val applicationFilesDir: String, private val resources: Resources) {
-    fun findIconUri(appName: String): Uri {
-        val icon =
-                File("$applicationFilesDir/apps/$appName/$appName.png")
-        if (icon.exists()) return Uri.fromFile(icon)
-        return getDefaultIconUri()
-    }
-
-    private fun getDefaultIconUri(): Uri {
-        val resId = R.mipmap.ic_launcher_foreground
-        return Uri.parse(ContentResolver.SCHEME_ANDROID_RESOURCE +
-                "://" + resources.getResourcePackageName(resId) + '/' +
-                resources.getResourceTypeName(resId) + '/' +
-                resources.getResourceEntryName(resId))
-    }
-
-    fun findAppDescription(appName: String): String {
-        val appDescriptionFile =
-                File("$applicationFilesDir/apps/$appName/$appName.txt")
-        if (!appDescriptionFile.exists()) {
-            return resources.getString(R.string.error_app_description_not_found)
-        }
-        return appDescriptionFile.readText()
     }
 }
